Hybrid transport: Add state-assisted transactions (#100)
## Changes

* Updates internals of CableKnownDeviceInfoStore, to make it easier to
use and thread-safe.
* Adds parsing of caBLE update messages. This includes a Google-specific
key (999). Due to a limitation in serde-indexed, parsing of this
structure has to be done field-by-field with `serde_cbor`. To be fixed
properly with #66.
* Connect to known device from saved state.
* Verifies handshake signature. Note this currently only works for
linking information received on QR-initiated transactions. The protocol
doesn't seen to support subsequent updates, as the ephemeral QR-code
identity key is needed for verification.