Implement discoverable credential 'preferred' (#125)
Switched away from a context-less `From<MakeCredentialRequest> for
Ctap2MakeCredentialRequest` implementation to a dedicated function that
can additionally receive the device-info to make that transformation.
Now we also don't need to copy the original request and modify it (which
was always a bit icky), but we can do the direct translation to the
Ctap2-request.