Support attested_credential and extensions both be present in AuthenticatorData
As described in #135, AuthenticatorData can contain both an attested_credential, as well as extensions when making a credential.
However, from_reader checks that there is no trailing data in the cbor of the public key by calling deserializer.end().
Therefore, we need to use the manual interface to deserialize.

Fixes #135.