Switch virt device to RAM only and use message passing to allow for state to be kept across device lifetime (#150)
### Commit 1:
Since the callback structure of nitrokeys/trusseds firmware makes it
hard to have a 'normal' device, that can handle more than one
consecutive requests without losing state, and after banging my head
against this for quite some time, I opted for a rather blunt, but
hopefully effective (enough) solution: Simple message passing.

Now the callback-function is started in its own thread and
requests+responses are passed back and forth via normal channels.
I also removed the filesystem storage and went back to RAM storage only.
This way, it should be possible to have multiple devices in parallel
without any interference. As long as the device object lives, state is
kept and things like "establishing shared secret, getting
pinUvAuthToken, then doing the request" work, as the device is not reset
between each request.

I added a rudimentary test-battery for PRF, demonstrating that this can
now be done.

I also added a few more dev-dependencies to allow for better
usage/debugging of this virtual device. With these, it is possible to
show logging from the virtual fido-authenticator in the end2end-tests.

### Commit 2:
Renaming `TestChannel` to `MockChannel` (which I think explains better
what it is) and moving the file out of `src/tests/`. I realized this
would be very confusing, as in there are only tests that use the virtual
device and not the `MockChannel`.
So now we have two testing transports under `src/transports/`, namely
`virt` and `mock`, which hopefully is a bit more self-explanatory.

### Commit 3:
Add tests for preflight, to show how testing individual functions could look like with the virtual device

### Commit 4: 
Also test UvUpdate-calls in PRF tests. I found two bugs while doing that:
    PresenceRequired was not sent, if certain pinUvAuth-methods were used.
    PresenceRequired was sometimes sent, even if the platform errors out when calculating HMACs.