Log when CTAP2 bio_enrollment responses return None fields (#153)
This PR adds `info!` logs in the error paths for a couple of
`BioEnrollment` implementation methods, which previously returned
`Error::Ctap(CtapError::Other)` silently under some CTAP requests to
channels, where the response had `None` fields that were expected to be
`Some`.

Maintainers: Are these logs desirable in the context of #14 ? How
appropriate are the messages and log-levels I have chosen?